requirements.txt
google-adk>=0.6.0
anthropic>=0.40.0
litellm>=1.50.0
google-cloud-aiplatform>=1.70.0
vertexai>=1.70.0
pydantic>=2.0.0
python-dotenv>=1.0.0
